Tests/LibWeb/Text/expected/abspos-relayout-continuation.txt

rects-match=true
partial-layouts=0 full-layouts=1
FAIL
